requirements.txt
boto3==1.33.3
python-dotenv==1.0.0
streamlit==1.31.0